split into publicApi
