+++ index.js
@@ -1,2 +1,3 @@
 module.exports.tradingApi = require('./tradingApi');
-module.exports.streamApi = require('./streamApi');
+module.exports.publicApi = require('./publicApi');
+module.exports.pushApi = require('./pushApi');
